refactor ProgressiveSolveHelper
